- add `IdSegmentDistributorSpec` - rename `MachineIdDistributorSpec`
