+++ cosid-test/src/main/java/me/ahoo/cosid/test/segment/distributor/IdSegmentDistributorSpec.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package me.ahoo.cosid.test.segment.distributor;
+
+
+import me.ahoo.cosid.segment.IdSegmentDistributorFactory;
+
+/**
+ * IdSegmentDistributorSpec .
+ * TODO
+ *
+ * @author ahoo wang
+ */
+public abstract class IdSegmentDistributorSpec {
+    
+    protected abstract IdSegmentDistributorFactory getFactory();
+    
+    public void getNamespace() {
+    
+    }
+    
+    public void getName() {
+    
+    }
+    
+    public void getStep() {
+    
+    }
+    
+    public void nextMaxId() {
+    
+    }
+}
