+++ gradle.properties
@@ -12,7 +12,7 @@
 #
 
 group=me.ahoo.cosid
-version=1.9.0-SNAPSHOT
+version=1.9.0-SNAPSHOT-2
 
 description=Universal, flexible, high-performance distributed ID generator.
 website=https://github.com/Ahoo-Wang/CosId
